refactor: Use enums to store cell states
Use `IntEnum` for states, move automaton to separate classes.
